fixed semantics of the mu parameter
following thompson (2005), mu = sqrt (1 - (rho/Rsun)^2) where rho is called
the heliocentric impact parameter. it is the length of the radius vector
from disc center to the target position.

In particular, mu is not the 'impact parameter' in terms of 'radial distance to
the solar rotation axis' as it was defined here previously...