re-enabled fits output; added elevation + airmass
 * most parameters can now be written to a fits binary table again
 * elevation and airmass estimates are added to the ephemeris parameters

The elevation is computed via h = 90 - z from the zenith angle z that is
computed via SPICE from the local normal to the surface. the elevation is
corrected for atmospheric refraction using the J. Meeus "Astronomical
Algorithms" formula. This needs a check, there seem to be some jumps in the
refraction corrections computed that way...